move CheckAccess
