+++ ydb/core/grpc_services/rpc_load_rows.cpp
@@ -109,6 +109,38 @@ bool ConvertArrowToYdbPrimitive(const arrow::DataType& type, Ydb::Type& toType) 
     return false;
 }
 
+bool CheckAccess(const TString& table, const TString& token, const NSchemeCache::TSchemeCacheNavigate* resolveResult, TString& errorMessage) {
+    if (token.empty())
+        return true;
+
+    NACLib::TUserToken userToken(token);
+    const ui32 access = NACLib::EAccessRights::UpdateRow;
+    if (!resolveResult) {
+        TStringStream explanation;
+        explanation << "Access denied for " << userToken.GetUserSID()
+                    << " table '" << table
+                    << "' has not been resolved yet";
+
+        errorMessage = explanation.Str();
+        return false;
+    }
+    for (const NSchemeCache::TSchemeCacheNavigate::TEntry& entry : resolveResult->ResultSet) {
+        if (entry.Status == NSchemeCache::TSchemeCacheNavigate::EStatus::Ok
+            && entry.SecurityObject != nullptr
+            && !entry.SecurityObject->CheckAccess(access, userToken))
+        {
+            TStringStream explanation;
+            explanation << "Access denied for " << userToken.GetUserSID()
+                        << " with access " << NACLib::AccessRightsToString(access)
+                        << " to table '" << table << "'";
+
+            errorMessage = explanation.Str();
+            return false;
+        }
+    }
+    return true;
+}
+
 }
 
 using TEvBulkUpsertRequest = TGrpcRequestOperationCall<Ydb::Table::BulkUpsertRequest,
@@ -176,7 +208,7 @@ private:
     }
 
     bool CheckAccess(TString& errorMessage) override {
-        return NTxProxy::CheckAccess(GetTable(), Request->GetSerializedToken(), GetResolveNameResult(), errorMessage);
+        return ::NKikimr::NGRpcService::CheckAccess(GetTable(), Request->GetSerializedToken(), GetResolveNameResult(), errorMessage);
     }
 
     TConclusion<TVector<std::pair<TString, Ydb::Type>>> GetRequestColumns() const override {
@@ -356,7 +388,7 @@ private:
     }
 
     bool CheckAccess(TString& errorMessage) override {
-        return NTxProxy::CheckAccess(GetTable(), Request->GetSerializedToken(), GetResolveNameResult(), errorMessage);
+        return ::NKikimr::NGRpcService::CheckAccess(GetTable(), Request->GetSerializedToken(), GetResolveNameResult(), errorMessage);
     }
 
     TConclusion<TVector<std::pair<TString, Ydb::Type>>> GetRequestColumns() const override {
